% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_irl_uk.R
\name{covid_mobility_irl_uk}
\alias{covid_mobility_irl_uk}
\title{IRL and UK Covid mobility data}
\source{
\href{https://github.com/joachim-gassen/tidycovid19}{tidycovid19} and \href{https://github.com/kjhealy/covdata}{covdata}
}
\usage{
covid_mobility_irl_uk()
}
\description{
Function that produces the covid mobility data for the major cities of IRL and UK
}
