% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_dt_tool.R
\name{dt_styler}
\alias{dt_styler}
\title{Datatable function}
\usage{
dt_styler(data, cnt1, cnt2, clr1, clr2)
}
\arguments{
\item{data}{this is the dataset where the information lies}

\item{cnt1}{this is country one}

\item{cnt2}{this is country two}

\item{clr1}{this is color one}

\item{clr2}{this is color two}
}
\value{
will return a datatable with colored rows according to the inputed country index
}
\description{
This function will build a datatable and color it for the inputted countries and color choices
}
