% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Generate the shiny application : \strong{Covid19: The Battle of Independence}}
\usage{
launch_app()
}
\description{
launch_app( ) builds and produces the shiny application \emph{Covid19: The Battle of Independence.} The application
displays a comparative and interactive analysis between the impact and response to covid19 between Ireland and the United Kingdom.
}
