% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\docType{data}
\name{mobility_major_cities}
\alias{mobility_major_cities}
\title{Mobility data for the major cities in Ireland and the United Kingdom}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9765 rows and 12 columns.
}
\source{
The data has been collected from \href{https://github.com/joachim-gassen/tidycovid19}{tidycovid19} package and the \href{https://github.com/kjhealy/covdata}{covdata} package
}
\usage{
mobility_major_cities
}
\description{
\code{mobility_major_cities} returns a dataset that contains the monility data for the major cities in Ireland and the United Kingdom with the variables listed below.
}
\author{
Samuel Lyubic
}
\keyword{datasets}
