% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_uk_irl.R
\name{covid_data_uk_irl}
\alias{covid_data_uk_irl}
\title{IRL and UK Covid data}
\source{
\href{https://github.com/joachim-gassen/tidycovid19}{tidycovid19} and \href{https://github.com/kjhealy/covdata}{covdata}
}
\usage{
covid_data_uk_irl()
}
\description{
Function that produces the covid data for IRL and UK
}
