% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_data_clean.R
\name{data_wrangle}
\alias{data_wrangle}
\title{A function that wrangles selected variables into proportions and long form}
\usage{
data_wrangle(data)
}
\arguments{
\item{data}{the dataset that is being used to generate the proportions and long form}
}
\description{
Takes the chosen variable and calculates a proportion and then turns it into long form
}
