% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_date_tool.R
\name{dateInput_range}
\alias{dateInput_range}
\title{Function that sets the date range for the slider}
\usage{
dateInput_range(inputId, label = "Chosen date range", x, start = min(x), ...)
}
\arguments{
\item{inputId}{input id}

\item{label}{is the label}

\item{x}{is the date data}

\item{start}{is the start date for the range}

\item{...}{is the rest of the inputs which are as follows}
}
\value{
will return a date range  with chosen date values as specified by the user
}
\description{
This function will allow the user to repeatedly add date slider inputs by adding the appropriate params
}
