% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_dt_tool.R
\name{dt_styler}
\alias{dt_styler}
\title{Datatable function}
\usage{
dt_styler(data, id, key1, key2, clr1, clr2, plength)
}
\arguments{
\item{data}{this is the dataset where the information lies}

\item{id}{is the variable column id where the variables are being targetted}

\item{key1}{this is country one}

\item{key2}{this is country two}

\item{clr1}{this is color one}

\item{clr2}{this is color two}

\item{plength}{this is the length of the table}
}
\value{
will return a datatable with colored rows according to the inputed country index
}
\description{
This function will build a datatable and color it for the inputted countries and color choices
}
