% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_covid_data.R
\docType{data}
\name{raw_covid_data_uk_irl}
\alias{raw_covid_data_uk_irl}
\title{Raw form of the covid data used for analysis in the application}
\format{
format is x by y
}
\usage{
raw_covid_data_uk_irl
}
\description{
covid data raw form
}
\author{
Samuel Lyubic
}
\keyword{datasets}
